reorganize files; clean up logic; split PerformanceEntry buffers
